Grand application redinement

+++ src/GenoDev.BusinessTracker.ApplicationLogic/UseCases/Production/GetProducts/GetProductsQuery.cs
@@ -18,5 +18,5 @@ public record GetProductsQuery(
     string? NameFilter = null,
     string? IdentifierFilter = null,
     string? DescriptionFilter = null,
-    int? AmountFilter = null,
+    double? AmountFilter = null,
     NumericOperator? AmountOperator = null) : IRequest<PagedList<ProductDto>>;
